LayoutTests/fast/element-targeting/target-container-with-anonymous-sibling.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<head>
<meta charset="utf-8">
<meta name="viewport" content="width=device-width, initial-scale=1">
<script src="../../resources/ui-helper.js"></script>
<script src="../../resources/js-test.js"></script>
<style>
body, html {
    margin: 0;
}

.box {
    width: 100px;
    height: 100px;
    background: tomato;
    position: fixed;
    top: 0;
    left: 0;
    z-index: 100;
}
</style>
</head>
<body>
Here’s to the crazy ones. The misfits. The rebels. The troublemakers. The round pegs in the square holes. The ones who see things differently. They’re not fond of rules. And they have no respect for the status quo.
<div class="box"></div>
</body>
<script>
jsTestIsAsync = true;

addEventListener("load", async event => {
    box = document.querySelector(".box");
    selector = await UIHelper.adjustVisibilityForFrontmostTarget(box);
    shouldBe("box", "document.querySelector(selector)");
    finishJSTest();
});
</script>
</html>
