LayoutTests/fast/element-targeting/target-container-with-anonymous-sibling-expected.txt

PASS box is document.querySelector(selector)
PASS successfullyParsed is true

TEST COMPLETE
Here’s to the crazy ones. The misfits. The rebels. The troublemakers. The round pegs in the square holes. The ones who see things differently. They’re not fond of rules. And they have no respect for the status quo.
